More comment searches
People link their steam community page as opposed to just leaving the
name. This confusing way of doing things messes up the bot so the bot
will now search for people linking so as not to get confused.

(This may have the unintended side effect that people may link profiles
that are not their own, unintentionally, but since the bot checks for
those links first the scores will be misattributed. If this happens I'll
revert this change and make the previous score submission format more
clear)